+++ src/test/java/org/solid/testharness/reporting/ReportFragmentTest.java
@@ -321,9 +321,9 @@ class ReportFragmentTest {
         final String reportStripped = StringUtils.normalizeSpace(report);
         assertThat(reportStripped, matchesPattern(".*datetime=\"1970-01-01T00:00Z\".*"));
         assertThat(reportStripped, matchesPattern(".*<dd>[0-9]+ ms</dd>.*"));
-        assertTrue(reportStripped.contains("<td>All requirements (Features)</td> <td>1</td> " +
+        assertTrue(reportStripped.contains("<td>Totals</td> <td>1</td> " +
                 "<td>0</td> <td>0</td> <td>0</td> <td>0</td> <td>1</td>"));
-        assertTrue(reportStripped.contains("<td>All tests (Scenarios)</td> <td>1</td> <td>1</td> " +
+        assertTrue(reportStripped.contains("<td>Totals</td> <td>1</td> <td>1</td> " +
                 "<td>0</td> <td>0</td> <td>0</td> <td>2</td>"));
     }